# perf(storage): reduce allocations when accessing series file or TSI