--- a/tsdb/index_test.go
+++ b/tsdb/index_test.go
@@ -150,8 +150,7 @@ func (i *Index) Close() error {
 	if err := i.sfile.Close(); err != nil {
 		return err
 	}
-	//return os.RemoveAll(i.rootPath)
-	return nil
+	return os.RemoveAll(i.rootPath)
 }
 
 // This benchmark compares the TagSets implementation across index types.