This PR adds some performance improvements to the RHH structures used in the Series File and TSI index. These improvements will help with both reads and writes to the index. 

These improvements lead to upto a `98%` reduction in the number of objects allocated on the write path, and a `54%` reduction the amount of memory allocated. I would expect similar improvements on the read path, since the same APIs are called.

Previously all hashmap calls that are instrumented would create a new map (`prometheus.Labels`) whenever they were called, which caused a lot of allocations.

This PR pre-creates all the necessary maps (which are only used in a read-only capacity).

```
⇒  benchstat old.txt new.txt
name                                                     old time/op    new time/op    delta
Index_CreateSeriesListIfNotExist/create_series-8            5.00s ± 3%     5.13s ± 2%   +2.50%  (p=0.033 n=10+7)
Index_CreateSeriesListIfNotExist/already_exist_series-8     557ms ± 3%     530ms ± 6%   -4.85%  (p=0.001 n=10+8)

name                                                     old alloc/op   new alloc/op   delta
Index_CreateSeriesListIfNotExist/create_series-8           2.57GB ± 0%    1.84GB ± 1%  -28.52%  (p=0.000 n=8+10)
Index_CreateSeriesListIfNotExist/already_exist_series-8     678MB ± 0%     308MB ± 0%  -54.55%  (p=0.000 n=10+8)

name                                                     old allocs/op  new allocs/op  delta
Index_CreateSeriesListIfNotExist/create_series-8            28.9M ± 0%     24.5M ± 0%  -15.22%  (p=0.000 n=9+10)
Index_CreateSeriesListIfNotExist/already_exist_series-8     2.23M ± 0%     0.03M ± 0%  -98.51%  (p=0.000 n=10+10)
```